fix(sim): round creature, bonus, effect and camera math per x87 op
Both ports now follow the native x87 PC24 operation order, f32 literals and
store points in these paths:

- Quick Learner kill XP: __ftol(reward * 1.3f), not a double product
- Radioactive kill XP: exact fild of the experience, one PC24 fadd
- orbit_radius union: ranged projectile types are stored as int bits, so a
  recycled slot's radius/projectile type is read the way native reads it
- template 0x37 zeroes link_index
- spawner/long-strip movers clamp only before moving, against 1024 - size
- Telekinetic hover accumulates the int frame_dt_ms
- sprite effect fade, spin, growth and drift round per op (alpha 0.25 lasts
  16 ticks at 60 Hz, not 15)
- camera shake timer: PC24 decay and f32 resets (0.1f, 0.06f, Nuke 0.2f)
- creature anim phase: PC24 multiply chain in native order
- rush and survival spawns load elapsed time and experience exactly (fild),
  which matters past 2^24
- freeze/blood angles, shock burst rotation/scale, contact decal position
  (D3DXVec2Normalize) and the corpse-queue 0.001f stage

+++ tests/gameplay/test_effects.py
@@ -341,7 +341,7 @@ def test_particle_hit_deflects_rescales_spawns_fx_and_pushes_creature() -> None:
     assert particle.render_flag is False
     assert fx_queue.count == 1
     assert sprite_effects.entries[0].active
-    assert_float_close(sprite_effects.entries[0].color.a, 0.7)
+    assert_float_close(sprite_effects.entries[0].color.a, f32(0.7))
 
     deflect_step = f32(math.tau * 0.2)
     assert_float_close(float(particle.angle), deflect_step)
